Add new make target 'integration', have 'test' exclude them
The rationale for this new target is that for #28, we'd probably have
these tests depend on git being installed; and 'make test' should still
"just work" for everyone.

Also, this is differentiating test cases by their _names_, not labels.
This is because I haven't found a way to _exclude_ based on a label, to
have `make test` not run any of the integration tests. (Short of adding
label=unit to all the unit tests, which would of course be another
option.)

Signed-off-by: Stephan Renatus <srenatus@chef.io>